int is int32_t on 64 bit machines. who knew?

+++ AcidAnalyzer/drawNcurses.c
@@ -1,20 +1,20 @@
 #include <drawNcurses.h>
 
-void drawSpectrum(double *soundArray, int x_size, int y_size, int y_startingLine, bool uprigth)
+void drawSpectrum(double *soundArray, int64_t x_size, int64_t y_size, int64_t y_startingLine, bool uprigth)
 {
     static wchar_t *character[] = {L" ", L"\u2581" , L"\u2582" , L"\u2583" , L"\u2584" , L"\u2585" , L"\u2586" , L"\u2587" , L"\u2588"};
     if (uprigth)
     {
         color_set(1, NULL);
         // attroff(A_REVERSE);
-        for (int x = 0; x < x_size; x++)
+        for (int64_t x = 0; x < x_size; x++)
         {
-            int soundInt = 0;
+            int64_t soundInt = 0;
             if (soundArray[x] < -1000000000) soundInt = -1000000000;
             else soundInt = soundArray[x] * ((y_size) * 8) + 8;
-            for (int y = y_startingLine + y_size; y >= y_startingLine; y--)
+            for (int64_t y = y_startingLine + y_size; y >= y_startingLine; y--)
             {
-                int charIndex = soundInt;
+                int64_t charIndex = soundInt;
                 if (soundInt <= 0)
                 {
                     charIndex = 0;
@@ -33,14 +33,14 @@ void drawSpectrum(double *soundArray, int x_size, int y_size, int y_startingLine
     {
         // attron(A_REVERSE);
         color_set(2, NULL);
-        for (int x = 0; x < x_size; x++)
+        for (int64_t x = 0; x < x_size; x++)
         {
-            int soundInt = 0;
+            int64_t soundInt = 0;
             if (soundArray[x] < -1000000000) soundInt = -1000000000;
             else soundInt = soundArray[x] * ((y_size) * 8);
-            for (int y = y_startingLine; y < y_startingLine + y_size; y++)
+            for (int64_t y = y_startingLine; y < y_startingLine + y_size; y++)
             {
-                int charIndex = soundInt;
+                int64_t charIndex = soundInt;
                 if (soundInt <= 0)
                 {
                     charIndex = 8;
